# apps/api-go/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/instructors:
    get:
      operationId: InstructorsPublicController_findAll_v1
      parameters:
      - name: search
        required: false
        in: query
        description: '搜索: name/title/headline 模糊匹配'
        schema:
          type: string
      - name: expertiseIds
        required: false
        in: query
        description: 按专长 ID 过滤 (多选 OR)
        schema:
          type: array
          items:
            type: string
      - name: published
        required: false
        in: query
        description: null=全部, true=已发布, false=草稿
        schema:
          type: boolean
      - name: sort
        required: false
        in: query
        description: '排序: orderIndex | name | recent'
        schema:
          default: orderIndex
          type: string
      - name: page
        required: false
        in: query
        description: 页码 (1-based)
        schema:
          minimum: 1
          default: 1
          type: number
      - name: limit
        required: false
        in: query
        description: 每页数量, max 100
        schema:
          minimum: 1
          maximum: 100
          default: 24
          type: number
      responses:
        '200':
          description: ''
      summary: 讲师列表 (前台, 仅 published)
      tags:
      - instructors (public)
  /api/v1/instructors/{slug}:
    get:
      operationId: InstructorsPublicController_findBySlug_v1
      parameters:
      - name: slug
        required: true
        in: path
        description: 讲师 slug
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 讲师详情 (按 slug, 前台)
      tags:
      - instructors (public)
...
  /api/v1/admin/instructors/expertises:
    get:
      operationId: ExpertisesAdminController_findAll_v1
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 专长列表 (admin, 含禁用)
      tags:
      - instructor-expertises (admin)
    post:
      operationId: ExpertisesAdminController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpertiseDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 新建专长 (admin)
      tags:
      - instructor-expertises (admin)
  /api/v1/admin/instructors/expertises/{id}:
    patch:
      operationId: ExpertisesAdminController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpertiseDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 更新专长 (admin)
      tags:
      - instructor-expertises (admin)
    delete:
      operationId: ExpertisesAdminController_delete_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 删除专长 (admin, 级联删除关联 link)
      tags:
      - instructor-expertises (admin)
  /api/v1/admin/instructors:
    get:
      operationId: InstructorsAdminController_findAll_v1
      parameters:
      - name: search
        required: false
        in: query
        description: '搜索: name/title/headline 模糊匹配'
        schema:
          type: string
      - name: expertiseIds
        required: false
        in: query
        description: 按专长 ID 过滤 (多选 OR)
        schema:
          type: array
          items:
            type: string
      - name: published
        required: false
        in: query
        description: null=全部, true=已发布, false=草稿
        schema:
          type: boolean
      - name: sort
        required: false
        in: query
        description: '排序: orderIndex | name | recent'
        schema:
          default: orderIndex
          type: string
      - name: page
        required: false
        in: query
        description: 页码 (1-based)
        schema:
          minimum: 1
          default: 1
          type: number
      - name: limit
        required: false
        in: query
        description: 每页数量, max 100
        schema:
          minimum: 1
          maximum: 100
          default: 24
          type: number
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 讲师列表 (admin, 含草稿)
      tags:
      - instructors (admin)
    post:
      operationId: InstructorsAdminController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstructorDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 新建讲师 (admin)
      tags:
      - instructors (admin)
  /api/v1/admin/instructors/{id}:
    get:
      operationId: InstructorsAdminController_findOne_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 讲师详情 (admin, 按 id)
      tags:
      - instructors (admin)
    patch:
      operationId: InstructorsAdminController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstructorDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 更新讲师 (admin)
      tags:
      - instructors (admin)
    delete:
      operationId: InstructorsAdminController_softDelete_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 软删讲师 (admin, 置草稿 + 解除所有 link)
      tags:
      - instructors (admin)
  /api/v1/admin/instructors/reorder:
    post:
      operationId: InstructorsAdminController_reorder_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderInstructorsDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 拖拽排序讲师 (admin, 一次提交完整 ID 列表)
      tags:
      - instructors (admin)
  /api/v1/admin/courses/{courseId}/instructors:
    get:
      operationId: CourseInstructorsAdminController_list_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 获取该课程的所有讲师/导师 link
      tags:
      - course-instructors (admin)
    post:
      operationId: CourseInstructorsAdminController_link_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCourseDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 挂一个讲师/导师到课程 (admin)
      tags:
      - course-instructors (admin)
    put:
      operationId: CourseInstructorsAdminController_syncLinks_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncCourseLinksDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 整组同步课程的讲师/导师 link (admin, 一次 diff 覆盖)
      tags:
      - course-instructors (admin)
  /api/v1/admin/courses/{courseId}/instructors/{linkId}:
    delete:
      operationId: CourseInstructorsAdminController_unlink_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      - name: linkId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 解除一个讲师 link (admin)
      tags:
      - course-instructors (admin)
...
info:
  title: AI Academy API
  description: AI and LLM education platform API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateExpertiseDto:
      type: object
      properties:
        key:
          type: string
          description: 英文 key, 唯一, 程序内标识
          maxLength: 60
        label:
          type: string
          description: 中文 label
          maxLength: 80
        labelEn:
          type: string
          description: 英文 label
          maxLength: 80
        isActive:
          type: boolean
          description: 是否启用
          default: true
        orderIndex:
          type: number
          description: 排序权重
          default: 0
      required:
      - key
      - label
    UpdateExpertiseDto:
      type: object
      properties:
        key:
          type: string
          description: 英文 key, 唯一, 程序内标识
          maxLength: 60
        label:
          type: string
          description: 中文 label
          maxLength: 80
        labelEn:
          type: string
          description: 英文 label
          maxLength: 80
        isActive:
          type: boolean
          description: 是否启用
          default: true
        orderIndex:
          type: number
          description: 排序权重
          default: 0
    CreateInstructorDto:
      type: object
      properties:
        slug:
          type: string
          description: URL slug, 留空自动生成
          maxLength: 80
        name:
          type: string
          description: 中文主显示名
          maxLength: 80
        nameEn:
          type: string
          description: 英文/拼音别名
          maxLength: 80
        title:
          type: string
          description: 头衔 (自由文本, 不做 enum)
          maxLength: 120
        titleEn:
          type: string
          description: 英文头衔
          maxLength: 120
        headline:
          type: string
          description: 一句话标语 (摘要用)
          maxLength: 255
        headlineEn:
          type: string
          description: 英文标语
          maxLength: 255
        bio:
          type: string
          description: 长 bio, markdown
          maxLength: 5000
        bioEn:
          type: string
          description: 英文 bio
          maxLength: 5000
        avatarUrl:
          type: string
          description: 头像 URL
        company:
          type: string
          description: 当前就职公司
          maxLength: 120
        yearsOfExperience:
          type: number
          description: 从业年限
          minimum: 0
          maximum: 80
        linkedinUrl:
          type: string
          description: LinkedIn URL
        githubUrl:
          type: string
          description: GitHub URL
        twitterUrl:
          type: string
          description: Twitter/X URL
        websiteUrl:
          type: string
          description: 个人网站 URL
        contactEmail:
          type: string
          description: 联系邮箱 (admin 可见, 前台不暴露)
        notes:
          type: string
          description: admin 内部备注
          maxLength: 2000
        orderIndex:
          type: number
          description: 排序权重, 升序
          default: 0
        published:
          type: boolean
          description: true=发布, false/null=草稿
          default: false
        expertiseIds:
          description: 专长 ID 数组 (创建时一次性挂载)
          items:
            type: array
            items:
              type: string
          type: array
      required:
      - name
    UpdateInstructorDto:
      type: object
      properties:
        slug:
          type: string
          description: URL slug, 留空自动生成
          maxLength: 80
        name:
          type: string
          description: 中文主显示名
          maxLength: 80
        nameEn:
          type: string
          description: 英文/拼音别名
          maxLength: 80
        title:
          type: string
          description: 头衔 (自由文本, 不做 enum)
          maxLength: 120
        titleEn:
          type: string
          description: 英文头衔
          maxLength: 120
        headline:
          type: string
          description: 一句话标语 (摘要用)
          maxLength: 255
        headlineEn:
          type: string
          description: 英文标语
          maxLength: 255
        bio:
          type: string
          description: 长 bio, markdown
          maxLength: 5000
        bioEn:
          type: string
          description: 英文 bio
          maxLength: 5000
        avatarUrl:
          type: string
          description: 头像 URL
        company:
          type: string
          description: 当前就职公司
          maxLength: 120
        yearsOfExperience:
          type: number
          description: 从业年限
          minimum: 0
          maximum: 80
        linkedinUrl:
          type: string
          description: LinkedIn URL
        githubUrl:
          type: string
          description: GitHub URL
        twitterUrl:
          type: string
          description: Twitter/X URL
        websiteUrl:
          type: string
          description: 个人网站 URL
        contactEmail:
          type: string
          description: 联系邮箱 (admin 可见, 前台不暴露)
        notes:
          type: string
          description: admin 内部备注
          maxLength: 2000
        orderIndex:
          type: number
          description: 排序权重, 升序
          default: 0
        published:
          type: boolean
          description: true=发布, false/null=草稿
          default: false
        expertiseIds:
          description: 专长 ID 数组 (创建时一次性挂载)
          items:
            type: array
            items:
              type: string
          type: array
    ReorderInstructorsDto:
      type: object
      properties:
        orderedIds:
          description: 按新顺序排列的讲师 ID 列表
          type: array
          items:
            type: string
      required:
      - orderedIds
    LinkCourseDto:
      type: object
      properties:
        instructorId:
          type: string
          description: 讲师 ID
        role:
          type: string
          enum:
          - instructor
          - mentor
          description: '角色: instructor 主讲师 / mentor 导师'
        isPrimary:
          type: boolean
          description: 是否主讲师, 仅 instructor role 有效
          default: false
        orderIndex:
          type: number
          description: 展示顺序
          default: 0
      required:
      - instructorId
      - role
    SyncCourseLinksDto:
      type: object
      properties:
        links:
          description: 完整 link 列表, 后端 diff 同步
          type: array
          items:
            $ref: '#/components/schemas/LinkCourseDto'
      required:
      - links
...